requirements_seescans.txt
absl-py==2.1.0
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2024.2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.66.1
idna==3.10
imageio==2.9.0
importlib_metadata==8.5.0
Jinja2==3.1.3
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.0
numpy==1.18.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
pillow==10.2.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.6.3
six==1.16.0
sympy==1.12
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.1
torch==2.4.1  #+cu118
torchaudio==2.4.1   #+cu118
torchvision==0.19.1 #+cu118
tqdm==4.48.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
